Proofread DRAM AL operation of memory writing
In section 2.1.2, there are two seperate objects in the sentence,
"提高 AL 的電位一段足以讓電容充電或放電的時間", which are "提高
AL 的電位" and "一段足以讓電容充電或放電的時間". However, there
is no conjunction connecting them. Clarifying the relationship
between them is important for readers to understand the AL
operation of memory writing and capacitor. So I rewrite the
sentence to achieve this purpose.